requirements.txt
Flask
google-generativeai
pyswisseph